Move backend processing into an async task
The commit refactors the backend event processing pipeline to be an
async task rather than a function. We observed a discrepancy between
what the Foosball server posted to Slack, and what it showed on the
UI. This may have been because both the timer task and the event
task called event-state-handler at nearly the same point in time.

By moving the event processing into its own task, and using channels
for coordination, there is no way for two separate tasks to
influence the state at the same time. Now, if the tick task pushed
onto the channel before the event task, only one event would be
handled before the other. The channels are used for implicit queuing