Fix identical keys in time list loop
The previous implementation could result in keys that were identical
for goals scored within the same second. If the keys are identical,
the second timestamp would not show up in the list, and the Yavascript
throws an error. The commit updates the key generation to use a unique
key. Since there are only 9 elements, re-rendering the list isn't too
costly.